Add SourceMapper
That handles converting back and forth user-facing sources to
interactive compiler sources, and knows how to convert positions too.
