+++ metals/src/main/scala/scala/meta/internal/metals/MetalsLanguageServer.scala
@@ -264,6 +264,13 @@ class MetalsLanguageServer(
   private var testProvider: TestSuitesProvider = _
   private var workspaceReload: WorkspaceReload = _
   private var buildToolSelector: BuildToolSelector = _
+
+  private val sourceMapper = SourceMapper(
+    buildTargets,
+    buffers,
+    () => workspace
+  )
+
   def loadedPresentationCompilerCount(): Int =
     compilers.loadedPresentationCompilerCount()
   var tables: Tables = _
@@ -653,7 +660,6 @@ class MetalsLanguageServer(
             workspace,
             clientConfig,
             () => userConfig,
-            () => ammonite,
             buildTargets,
             buffers,
             symbolSearch,
@@ -664,7 +670,8 @@ class MetalsLanguageServer(
             excludedPackageHandler.isExcludedPackage,
             scalaVersionSelector,
             trees,
-            mtagsResolver
+            mtagsResolver,
+            sourceMapper
           )
         )
         debugProvider = new DebugProvider(
@@ -2316,7 +2323,8 @@ class MetalsLanguageServer(
     sh,
     symbolDocs,
     scalaVersionSelector,
-    mainBuildTargetsData
+    mainBuildTargetsData,
+    sourceMapper
   )
 
   private def checkRunningBloopVersion(bspServerVersion: String) = {
@@ -2481,16 +2489,7 @@ class MetalsLanguageServer(
         case NonFatal(e) =>
           scribe.error("unexpected error during source scanning", e)
       },
-      toIndexSource = path => {
-        if (path.isAmmoniteScript)
-          for {
-            targets <- buildTargets.sourceBuildTargets(path)
-            target <- targets.headOption
-            toIndex <- ammonite.generatedScalaPath(target, path)
-          } yield toIndex
-        else
-          None
-      }
+      toIndexSource = path => sourceMapper.mappedTo(path).getOrElse(path)
     )
   }
