+++ metals/src/main/scala/scala/meta/internal/metals/ammonite/Ammonite.scala
@@ -1,9 +1,7 @@
 package scala.meta.internal.metals.ammonite
 
 import java.io.InputStream
-import java.net.URI
 import java.nio.charset.Charset
-import java.nio.file.Paths
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
@@ -17,7 +15,6 @@ import scala.util.Failure
 import scala.util.Success
 import scala.util.control.NonFatal
 
-import scala.meta.inputs.Input
 import scala.meta.internal.bsp.BuildChange
 import scala.meta.internal.builds.BuildTools
 import scala.meta.internal.metals.MetalsEnrichments._
@@ -285,88 +282,11 @@ final class Ammonite(
     lastImportVersions = VersionsOption(None, None)
     disconnectOldBuildServer().asJavaObject
   }
-
-  def generatedScalaPath(
-      targetId: BuildTargetIdentifier,
-      source: AbsolutePath
-  ): Option[AbsolutePath] =
-    if (Ammonite.isAmmBuildTarget(targetId) && source.isAmmoniteScript)
-      buildTargets.scalaTarget(targetId).map { target =>
-        val rel = source.toRelative(workspace())
-        val path = Paths
-          .get(new URI(target.classDirectory))
-          .getParent
-          .resolve(
-            s"src/ammonite/$$file/${rel.toString.stripSuffix(".sc")}.scala"
-          )
-        AbsolutePath(path.toAbsolutePath.normalize)
-      }
-    else
-      None
-
-  def generatedScalaInputForPc(
-      targetId: BuildTargetIdentifier,
-      source: AbsolutePath
-  ): Option[(Input.VirtualFile, Position => Position)] =
-    generatedScalaPath(targetId, source)
-      .map { scalaPath =>
-        val scInput = source.toInputFromBuffers(buffers)
-        val input = {
-          val input0 = scalaPath.toInput
-          // ensuring the path ends with ".sc.scala" so that the PC has a way to know
-          // what we're giving it originates from an Ammonite script
-          input0.copy(path = input0.path.stripSuffix(".scala") + ".sc.scala")
-        }
-
-        /*
-
-        When given a script like
-
-            case class Bar(xs: Vector[String])
-
-        Ammonite generates a .scala file like
-
-            package ammonite
-            package $file
-
-            import _root_.ammonite.interp.api.InterpBridge.{value => interp}
-
-            object `main-1`{
-              /*<script>*/case class Bar(xs: Vector[String])/*</script>*/ /*<generated>*/
-              def $main() = { scala.Iterator[String]() }
-              override def toString = "main$minus1"
-              /*</generated>*/
-            }
-
-        When the script is being edited, we re-generate on-the-fly a valid .scala file ourselves
-        from the one originally generated by Ammonite. The result should be a valid scala file, that
-        we can pass to the PC.
-
-        In order to update the .scala file above, we:
-        - remove the section between '/*<generated>*/' and '/*</generated>*/'
-        - replace the section between '/*<script>*/' and '/*</script>*/' by the new content of the script
-
-         */
-
-        val updatedContent = input.value
-          .replaceAllBetween("/*<generated>*/", "/*</generated>*/")("")
-          .replaceAllBetween("/*<script>*/", "/*</script>*/")(
-            Ammonite.startTag + scInput.value
-          )
-        val updatedInput = input.copy(value = updatedContent)
-
-        val scriptStartIdx =
-          updatedContent.indexOf(Ammonite.startTag) + Ammonite.startTag.length
-        val addedLineCount = updatedContent.lineAtIndex(scriptStartIdx)
-        def updatedPos(position: Position) =
-          new Position(addedLineCount + position.getLine, position.getCharacter)
-        (updatedInput, updatedPos)
-      }
 }
 
 object Ammonite {
 
-  private def startTag: String =
+  def startTag: String =
     "/*<start>*/\n"
 
   case class AmmoniteMetalsException(msg: String) extends Exception(msg)
